add total contrib some with optional flag
